feat(templates): Add support for managing templates from admin console
Instead of relying on templates on disk on the backend server, templates
have been moved into the database. Here we're adding support for
creating, editing and removing templates, so users are able to customize
notification messages.

There is also an option for allow users to re-import templates from disk
to override any local changes that may have been made by the user.